How a computer can invent by itself
Methods for developing inventions, with the help of which three programmers can easily create a program using which a computer can invent many inventions by itself

Hello. IBM and Softline companies help me to implement the work called: How a computer can invent by itself (i.e. the Methods for developing inventions, with the help of which three programmers can easily create a program using which a computer can invent many inventions by itself). I believe that copies of this program can be sold for hundreds of millions of dollars. I ask you to contribute to implementation (i.e. use) of this work, or I ask you to implement (i.e. use) this work. I am the author of this work. This work and my phone are set out on the website http://www.55255.ru/ I believe that with the help of my abovementioned work the «Method» company has created the program using this program a computer can invent by itself a lot inventions. More about this program is stated on the website http://www.method.ru/ But the creators of this program apparently has not published information that they have used (I suppose) my abovementioned work to create this program. I believe that with the help of my abovementioned work the «Predizo» company has created the program (this program is called «True Machina») using this program a computer can invent by itself a lot inventions. More about program «True Machina» is stated on the website https://www.truemachina.com/ But the creators of the «True Machina» program apparently has not published information that they have used (I suppose) my abovementioned work to create this program. So I am striving for creation for the third time and with the help of this my work of the program using which a computer could independently invent many inventions. My e-mail 275527@gmail.com
The business plan: I am the Director of «Tonna zolota» LLC. I'll hire three programmers. And in one year they will easily create a program with the help of this work, by means of which a computer will be able to invent many inventions by itself. For the salary of programmers and other expenses $ 50000 will be necessary. 49% of profit on sale of inventions, program copies (and the other) will be yours, and I will take 51%. To implement this plan it is necessary that you give me $ 50000 or hire three programmers yourself.
Yours faithfully, Shmonov Aleksandr




Here is a summary of the work that has the title: How a computer can invent by itself (i.e. the Methods for developing inventions with the help of which three programmers can easily create a program using which a computer can invent many inventions by itself)
 
Let’s suppose that two such conditional propositions are written to the computer memory (and also other conditional propositions are written):
1) If: fire is placed under the stone, then: the stone will heat up.
2) If: the stone will heat up, then: the stone will expand.
Words of conditional proposition which stand from (i.e. after) the word «if» and before the word «then» are called the basis of conditional proposition, and words of conditional proposition that stand after the word «then» are called the consequence of conditional proposition.
Let’s suppose that computer should solve the following inventive task, i.e. the computer has to determine what needs to be done to have the following: the stone will expand (i.e. the computer has to determine how the following can be obtained: the stone will expand), let’s call this task the original inventive task (let’s assume that this task has not been solved yet). From the second conditional proposition it follows that in order for the computer to solve the original inventive task it is necessary for the computer to solve the following inventive task, i.e. it is necessary for the computer to determine what needs to be done to obtain the following: the stone will heat up (i.e. it is necessary for the computer to determine how the following can be obtained: the stone will be heated); let’s call this task the second inventive task. And (from the first conditional proposition it follows that) in order for the computer to solve the second inventive task, it is necessary for it to solve the following inventive task, i.e. it is necessary for the computer to determine what needs to be done to have the following: fire will be placed under a stone (let's call this problem the third inventive task). ))And the third inventive task has been solved, because it is known how to get the following: fire will be placed under a stone. And if the third inventive task has been solved, then the second inventive task has been solved too. And if the second inventive task has been solved, then the original inventive task has been solved too.
The Rule: Let’s take any inventive task (let's call this inventive task the fourth inventive task). In order for a computer to create an inventive task, having solved which it thereby solved the fourth inventive task, it is necessary for the computer to find in its own memory such a conditional proposition that has the following feature: the consequence of this conditional proposition and description of this fourth inventive task have the same meanings or consist of the same words which are located in the same sequence. And the basis of this conditional proposition will be an inventive task, having solved which the computer thereby solves the fourth inventive task. They have the same meanings: a) the word and interpretation of this the word b) synonyms and so on.
 Computer can find the same words in its memory. Let's take any inventive task (let's call this inventive task the fifth inventive task). The computer will solve the fifth inventive task if it does the following: first, using this rule, it will create such an inventive task (let’s call this task the sixth inventive task), having solved which it thereby solves the fifth inventive task, then, using this rule, the computer will create such an inventive task, having solved which it thereby  solved the sixth inventive task, etc., (on average 90 times) to the moment at which (i.e. until) the computer creates such an inventive task the solution of which is known, and if the computer creates such (i.e. the latter) inventive task, then the computer will solve the fifth inventive task. That is, the computer will solved the fifth (i.e. any) inventive task if it creates on average 90 such tasks.
       Almost all currently known information (which is needed to create inventions) can be expressed in the form of conditional propositions. If, for example, 400 random physical effects in the form of conditional propositions are stored in the computer memory, then the computer can create on average a lot of inventions using this method (an average inventor knows 150 physical effects).